rename: rename common to global
