+++ src/main/java/com/commerce/backendserver/global/exception/ApiExceptionHandler.java
@@ -1,7 +1,7 @@
-package com.commerce.backendserver.common.exception;
+package com.commerce.backendserver.global.exception;
 
-import com.commerce.backendserver.common.exception.error.ErrorCode;
-import com.commerce.backendserver.common.exception.error.ErrorResponse;
+import com.commerce.backendserver.global.exception.error.ErrorCode;
+import com.commerce.backendserver.global.exception.error.ErrorResponse;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
