+++ src/main/java/com/commerce/backendserver/global/exception/error/GlobalError.java
@@ -1,4 +1,4 @@
-package com.commerce.backendserver.common.exception.error;
+package com.commerce.backendserver.global.exception.error;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
